son düzenlemeler

+++ Story.Domain/Entities/Chapter.cs
@@ -9,10 +9,10 @@ namespace Story.Domain.Entities
 {
     public class Chapter
     {
-        public int Id { get; set; }
-        public int StoryId { get; set; }
+        public short Id { get; set; }
+        public short StoryId { get; set; }
         public string Title { get; set; }
-        public int Order { get; set; }                // Bölüm sıralaması
+        public short Order { get; set; }                // Bölüm sıralaması
         public bool IsCompleted { get; set; }         // Yayınlanmış versiyonu var mı
         public DateTime CreatedAt { get; set; }
         public string? Summary { get; set; }          // Bölüm özeti veya açıklaması
